# docs/swagger.yaml
basePath: /
definitions:
...
  modelsuser.RoleType:
    enum:
    - admin
    - operator
    - accountant
    type: string
    x-enum-varnames:
    - AdminRole
    - OperatorRole
    - AccountantRole
  modelsuser.User:
    type: object
  modelsuser.UserRes:
    properties:
      firstname:
        example: John
        type: string
      id:
        example: 1
        format: int64
        type: integer
      isActive:
        example: true
        type: boolean
      lastname:
        example: Doe
        type: string
      park_no:
        example: P123
        type: string
      role:
        allOf:
        - $ref: '#/definitions/modelsuser.RoleType'
        enum:
        - admin
        - operator
        - accountant
        example: operator
      username:
        example: johndoe
        type: string
    type: object
...
host: 192.168.100.7:3000
info:
  contact: {}
  title: Airline REST API
paths:
...
  /api/v1/user/operators:
    get:
      consumes:
      - application/json
      description: Retrieves a list of users who have the role "operator" in descending
        order by ID
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of operators with pagination metadata
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No operators found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error retrieving users with operator role
          schema:
            additionalProperties: true
            type: object
      summary: Get all operators
      tags:
      - Users
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Retrieves a list of users with pagination support
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Can not retrieve users
          schema:
            type: string
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user in the database
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/modelsuser.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsuser.User'
        "400":
          description: Can not parse
          schema:
            type: string
        "500":
          description: Can not create
          schema:
            type: string
      summary: Create a new user
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Deletes a user's information from the database using their unique
        ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error deleting user
          schema:
            type: string
      summary: Delete a user by ID
      tags:
      - Users
    get:
      description: Retrieves a user from the database using their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsuser.UserRes'
      summary: Get user by ID
      tags:
      - Users
    put:
      description: Updates a user's data (isActive, username, firstname, lastname,
        etc.) in the database based on the input provided
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/modelsuser.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsuser.UserRes'
        "400":
          description: Invalid user data
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error updating user
          schema:
            type: string
      summary: Update user fields based on the provided data
      tags:
      - Users
swagger: "2.0"